Editing ft_putnbr decla for long long

+++ mylibc/multift.c
@@ -6,7 +6,7 @@
 /*   By: e.diabriek <e.diabriek@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/08 04:47:25 by evil.diabri       #+#    #+#             */
-/*   Updated: 2026/07/09 18:27:29 by e.diabriek       ###   ########.fr       */
+/*   Updated: 2026/07/10 14:36:03 by e.diabriek       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	ft_putchar(char c)
 	write(1, &c, 1);
 }
 
-void	ft_putnbr(int integer)
+void	ft_putnbr(long long integer)
 {
 	if (integer < 0)
 	{
