+++ mylibc/multift.h
@@ -6,12 +6,12 @@
 /*   By: e.diabriek <e.diabriek@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/08 04:02:05 by evil.diabri       #+#    #+#             */
-/*   Updated: 2026/07/09 18:17:54 by e.diabriek       ###   ########.fr       */
+/*   Updated: 2026/07/10 14:35:45 by e.diabriek       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 void	ft_putchar(char c);
-void	ft_putnbr(int n);
+void	ft_putnbr(long long n);
 void	ft_putstr(char *str);
 void	ft_swap(int *a, int *b);
 void	ft_rev_int_tab(int *tab, int size);
